+++ src/main/java/com/moa/global/config/db/JpaConfiguration.java
@@ -1,4 +1,4 @@
-package com.moa.user.config.db;
+package com.moa.global.config.db;
 
 
 import org.springframework.beans.factory.annotation.Qualifier;
